fixes
Move thresholdkey from LoginView to ThresholdKeyView, pass userdata instead.
Group storage_layer variable with respective threshold_key variable instead of being global variable to contentview.
Fix export share button to only be available after reconstruction instead of before initialization.
Split views into separate files and name appropriately.
Remove unnecessary code.
Remove commented code.
Remove unnecessary file creation comments.
