+++ Shared/LoaderView.swift
@@ -1,0 +1,12 @@
+import SwiftUI
+
+struct LoaderView: View {
+    var tintColor: Color = .blue
+    var scaleSize: CGFloat = 1.0
+
+    var body: some View {
+        ProgressView()
+            .scaleEffect(scaleSize, anchor: .center)
+            .progressViewStyle(CircularProgressViewStyle(tint: tintColor))
+    }
+}
