+++ Shared/LoginModel.swift
@@ -1,0 +1,40 @@
+import Foundation
+import CustomAuth
+
+class LoginModel: ObservableObject {
+    @Published var loggedIn: Bool = false
+    @Published var isLoading = false
+    @Published var navigationTitle: String = ""
+    @Published var userData: [String: Any]!
+
+    func setup() async {
+        await MainActor.run(body: {
+            isLoading = true
+            navigationTitle = "Loading"
+        })
+        await MainActor.run(body: {
+            if self.userData != nil {
+                loggedIn = true
+            }
+            isLoading = false
+            navigationTitle = loggedIn ? "UserInfo" : "SignIn"
+        })
+    }
+    func loginWithAuth0() {
+        Task {
+            let sub = SubVerifierDetails(loginType: .web,
+                                         loginProvider: .google,
+                                         clientId: "221898609709-obfn3p63741l5333093430j3qeiinaa8.apps.googleusercontent.com",
+                                         verifier: "google-lrc",
+                                         redirectURL: "tdsdk://tdsdk/oauthCallback",
+                                         browserRedirectURL: "https://scripts.toruswallet.io/redirect.html")
+            let tdsdk = CustomAuth(aggregateVerifierType: .singleLogin, aggregateVerifier: "google-lrc", subVerifierDetails: [sub], network: .TESTNET)
+            let data = try! await tdsdk.triggerLogin()
+            await MainActor.run(body: {
+                self.userData = data
+                loggedIn = true
+            })
+        }
+    }
+
+}
