+++ Shared/tkey_iosApp.swift
@@ -1,17 +1,10 @@
-//
-//  tkey_iosApp.swift
-//  Shared
-//
-//  Created by David Main on 2022/08/29.
-//
-
 import SwiftUI
 
 @main
 struct tkey_iosApp: App {
     var body: some Scene {
         WindowGroup {
-            ContentView(vm: ViewModel())
+            ContentView(vm: LoginModel())
         }
     }
 }
