+++ Tests iOS/Tests_iOS.swift
@@ -1,10 +1,3 @@
-//
-//  Tests_iOS.swift
-//  Tests iOS
-//
-//  Created by David Main on 2022/08/29.
-//
-
 import XCTest
 import tkey_ios
