+++ Tests iOS/Tests_iOSLaunchTests.swift
@@ -1,10 +1,3 @@
-//
-//  Tests_iOSLaunchTests.swift
-//  Tests iOS
-//
-//  Created by David Main on 2022/08/29.
-//
-
 import XCTest
 
 class Tests_iOSLaunchTests: XCTestCase {
